requirements.txt
pandas==2.3.1
Flask==3.1.1
Flask-CORS==6.0.1
numpy==2.3.2
exceptiongroup==1.3.0
earthaccess==0.14.0
dotenv==0.9.9
datetime==5.5
geojson==3.2.0
bs4==0.0.2
geopy==2.4.1
requests_cache==1.2.1
retry_requests==2.0.0
openmeteo_requests==1.7.0
jwt==1.4.0
gunicorn==20.1.0
PyJWT>=2.10.1,<3.0.0
supabase==2.18.1
supabase-auth>=2.12.3,<2.13.0